pneumonia_models/requirements.txt
tensorflow==2.15.0
flask==3.0.3
flask-cors==4.0.1
numpy==1.26.4
Pillow==10.3.0
opencv-python-headless==4.9.0.80
gunicorn==22.0.0
scikit-learn==1.4.2
